Fix AD7705 driver bugs found while processing PR #163
- Go: fix chip package import path (adcdac -> adc_dac dir), buffered
  bool parameter shadowing the buffered register-bit constant, and
  TinyGo machine.SPIConfig field names (MOSI/MISO -> SDO/SDI); add
  the resetPin support Go's own RFM9x driver already established as
  convention, closing a gap versus the other 6 language targets
- Kotlin: mark non-const companion members @JvmStatic so AD7705Full
  can access them (protected static access from a subclass companion
  is a known unsupported case otherwise)
- Groovy: fix `? CH1 else CH2` (invalid Groovy ternary syntax)
- C++: add missing cpp/examples/linux/adc_dac/AD7705/ (all three
  tiers) - the only language target's host example was missing
  entirely; add `using AD7705Minimal::read_raw/read_voltage;` so
  the base overloads aren't hidden on AD7705Full; fix nanosleep()
  needing <ctime>; fix CMakeLists.txt CPP_DIR path depth for the
  ESP-IDF/Pico SDK/Zephyr test apps; fix Zephyr test app referencing
  main.cpp instead of src/main.cpp; fix Pico SDK spi_set_format()
  integer args instead of SPI_CPOL_1/SPI_CPHA_1; fix printf %u vs
  uint32_t format mismatch on 32-bit embedded targets; replace
  tautological uint16_t <= 65535 test assertions (dead code, and a
  hard error under ESP-IDF's -Werror=type-limits); make the
  Pico-SDK/Zephyr delay fallback build without libc (some minimal-libc
  configs lack <chrono>/<ctime>)
- Regenerate cpp/README.md and the two Node.js package READMEs
- Regenerate rust/Cargo.lock for the new AD7705 crates

+++ cpp/examples/espidf/adc_dac/AD7705/complete/main/main.cpp
@@ -43,7 +43,7 @@ extern "C" void app_main(void) {
 
     uint32_t off2 = adc.get_offset_calibration(2);                                  // Read offset calibration, (channel=2) → uint32_t 24-bit
     uint32_t gain2 = adc.get_gain_calibration(2);                                   // Read gain calibration, (channel=2) → uint32_t 24-bit
-    printf("ch2 offset=%u gain=%u\n", off2, gain2);
+    printf("ch2 offset=%lu gain=%lu\n", (unsigned long)off2, (unsigned long)gain2);
 
     uint16_t raw1 = adc.read_raw(1);                                                // Read raw 16-bit code, (channel=1) → uint16_t
     float v1 = adc.read_voltage(1);                                                 // Read voltage, (channel=1) → float V
